Revert "refactored product"
This reverts commit 82b666767632e28a4c280dbec81d5acac72571bc.

+++ backend/superprice/src/main/java/com/example/superprice/repositories/SuperpriceRepositoryImpl.java
@@ -22,7 +22,7 @@ public class SuperpriceRepositoryImpl implements SuperpriceRepository {
 
     private Product extractProduct(ResultSet rs) throws SQLException {
         return new Product(rs.getLong(1), rs.getString(2), rs.getString(3),
-                rs.getString(4), rs.getString(5), rs.getDouble(6), rs.getInt(7));
+                rs.getString(4), rs.getString(5), rs.getLong(6), rs.getInt(7));
     }
 
     @Override
